+++ ChangeLog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added support for Ubuntu `24.04` and dropped support for Ubuntu `18.04`
 
 ## 1.6.3 - 2024-11-29